contact

+++ backend/controllers/contactController.js
@@ -1,42 +1,51 @@
 const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 
-
 //@description Update Contact
-//@route Put /api/contacts/:id
+//@route PUT /api/contacts/:id
 //access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
-    if(!contact) {
+    if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
     }
 
-    const updateContact = await Contact.findByIdAndUpdate(
+    if (contact._id.toString() !== req.contact.id) {
+        res.status(403);
+        throw new Error("You are not allowed to update this contact");
+    }
+
+    const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
-        {new: true}
+        { new: true }
     );
 
-    res.status(200).json(updateContact);
+    res.status(200).json(updatedContact);
 });
 
 //@description Delete Contact
-//@route Delete /api/contacts/:id
+//@route DELETE /api/contacts/:id
 //access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
-    if(!contact) {
+    if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+
+    if (contact._id.toString() !== req.contact.id) {
+        res.status(403);
+        throw new Error("You are not allowed to delete this contact");
     }
 
     await Contact.findByIdAndDelete(req.params.id);
 
-    res.status(200).json(contact);
+    res.status(200).json({ message: "Contact deleted successfully" });
 });
 
 module.exports = {
     updateContact,
-    deleteContact
-}
+    deleteContact,
+};
